requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1
backoff>=2.2.1
PyYAML>=6.0
numpy>=1.22.4
python-dateutil==2.8.2
pytz==2023.3
tzdata>=2022.7
urllib3>=1.26
trio>=0.17
trio-websocket>=0.9
certifi>=2021.10.8
typing_extensions>=4.9
websocket-client>=1.8
streamlit==1.31.0
aiohttp==3.9.1
asyncio==3.4.3
tqdm==4.66.1
